fix sonar issues

+++ can-lite/client/CanProtocolClient.cpp
@@ -59,7 +59,7 @@ namespace services
 
     void CanProtocolClient::DispatchPdu(uint32_t rawId, infra::ConstByteRange pdu)
     {
-        auto sourceNodeId = static_cast<uint16_t>(ExtractCanNodeId(rawId));
+        auto sourceNodeId = ExtractCanNodeId(rawId);
         if (sourceNodeId != 0)
             MarkServerAlive(sourceNodeId);
